LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/fetch/http-cache/304-update-expected.txt


PASS HTTP cache updates returned headers from a Last-Modified 304. 
FAIL HTTP cache updates stored headers from a Last-Modified 304. assert_less_than: Response used expected a number less than 3 but got 3
PASS HTTP cache updates returned headers from a ETag 304. 
FAIL HTTP cache updates stored headers from a ETag 304. assert_less_than: Response used expected a number less than 3 but got 3
